Refactoring routes, so you can work each route in its own file to avoid conflicts
