+++ app.js
@@ -5,6 +5,8 @@ const views = require('koa-views');
 const serve = require('koa-static');
 const config = require('./config');
 const db = require('./db');
+const fs = require('fs');
+
 // This library help us to interact with post requests
 const parse = require('co-body');
 
@@ -16,73 +18,10 @@ app.use(views('views', {
   map: config.templates
 }));
 
-function *home(){
-  yield this.render('index', {});
-}
-
-function *about_us(){
-
-  yield this.render('about_us', {});
-}
-
-function *contact(){
-  yield this.render('contact', {});
-}
-
-function *donations(){
-  yield this.render('donations', {});
-}
-
-function *events(){
-  yield this.render('events', {});
-}
-
-function *speakers(){
-  var  context =  {'usergroups': yield db.usergroups.find({})};
-
-  if(this.method === 'GET'){
-    yield this.render('speakers', context);
-  }
-
- if(this.method === 'POST'){
-    var talk = yield parse(this);
-
-    if(talk.usergroups.length > 0){
-
-      talks.insert(talk, function(err){
-        if(err) throw err;
-      });
-
-      context.talk = talk;
-      context.success = true;
-      console.log(talk);
-      yield this.render('speakers', context);
-
-    }
-
-    if( talk.usergroups.lenght === 0){
-      yield this.render('speakers', context);
-    }
-
-  }
-
-}
-
-function *usergroup(){
- var group =  new db.usergroups({
-    name: "Python Dominicana",
-    facebook_url: "https://www.facebook.com/groups/pythondo/",
-    website: "www.python.com.do",
-    logo: "https://scontent-mia1-1.xx.fbcdn.net/hphotos-ash2/v/t1.0-9/10624973_10204014303853300_9205003069849326284_n.jpg?oh=7bcdfcda40ccb89341d73ebfe9bcdbb6&oe=570E52D0",
-    created: new Date("2013-09-25T19:00:00.001Z"),
-    description: "Python Dominicana is an amazing place, is the place where Programmers, engineers an all kind of people interested in the Python Computer Language get together to talk, learn and share about all the stuff that can be done with it, from web applications to robots, from distributed systems to desktop apps.",
-  });
-  group.save(function(err){
-       if(err) throw err
-  });
-
-  this.body = "OK";
-}
+fs.readdirSync('./routes/').forEach(function(file) {
+    var route='./routes/'+file;
+    require(route)(app,db,routes);
+});
 
 function *users(){
 
@@ -103,17 +42,5 @@ function *pageNotFound(){
   }
 }
 
-
-app.use(routes.get('/', home));
-app.use(routes.get('/about', about_us));
-app.use(routes.get('/contact', contact));
-app.use(routes.get('/donations', donations));
-app.use(routes.get('/events', events));
-app.use(routes.get('/speakers', speakers));
-app.use(routes.post('/speakers', speakers));
-app.use(routes.get('/usergroup', usergroup));
 app.use(pageNotFound);
-
-
 app.listen(config.port);
-
