+++ routes/contact.js
@@ -1,0 +1,5 @@
+module.exports = function (app, db, routes){
+  app.use(routes.get('/contact', function *(){
+    yield this.render('contact', {});
+  }));
+};
