+++ routes/events.js
@@ -1,0 +1,5 @@
+module.exports = function (app, db, routes){
+  app.use(routes.get('/events', function *(){
+    yield this.render('events', {});
+  }));
+};
